Adds game logic to tic-tac-toe game
I created a gameLogic.js file, exporting three functions that
handle 1) placing Xs and Os when player clicks on board,
2) checking win conditions and ending play if met and 3) rendering
the game board.  One bug that remains is the draw counter, which
increases by 2 (instead of 1) every time there is a draw, and continues
to increase with clicks even after the game ends.  There is also a
console error, which doesn't break the game, but needs to be resolved.
The console error has to do with the gameBoard array being undefined
before play starts