Improves UX by hiding/un-hiding buttons and content
Previously all buttons showed all the time, and all
content showed all the time. Now, only the buttons a
user should click are showing, and the stay of the main
display provides prompting. I also addeded error messages
to the input modals, with the exception of the change password
model, which I need to do next.