Rebuilds game logic to use 'cells' array
Previously my game logic operated on a mirror local array,
instead of testing win conditions using the 'cells' array
handed back by the AJAX patch request. I changed so much in
such a short period of time, and the result is very messy,
but I learned a lot in the process, and I can continue to
refactor.